requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
farm-haystack
filelock==3.13.1
Flask==3.0.2
Flask-Session==0.8.0
fsspec==2024.3.1
huggingface-hub==0.21.4
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
packaging==24.0
PyMuPDFb==1.24.0
PyPDF2==3.0.1
PyYAML==6.0.1
redis==5.0.3
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sentence-transformers==2.2.2
sympy==1.12
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.18.1

